HTBHF-115 Adding new 404 error test.
